requirements.txt
-i https://pypi.org/simple
affine==2.3.1
altair==4.2.0; python_version >= '3.7'
attrs==22.1.0; python_version >= '3.5'
backports.zoneinfo==0.2.1; python_version < '3.9'
blinker==1.5; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
boto3==1.24.73
botocore==1.27.73; python_version >= '3.7'
cachetools==5.2.0; python_version ~= '3.7'
certifi==2022.9.14; python_version >= '3.6'
cffi==1.15.1
cfgrib==0.9.10.1
charset-normalizer==2.1.1; python_version >= '3.6'
click==8.1.3; python_version >= '3.7'
click-plugins==1.1.1
cligj==0.7.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2' and python_version < '4'
commonmark==0.9.1
cycler==0.11.0; python_version >= '3.6'
decorator==5.1.1; python_version >= '3.5'
eccodes==1.3.1
ecmwflibs==0.4.17
entrypoints==0.4; python_version >= '3.6'
findlibs==0.0.2
fiona==1.8.21
flatbuffers==2.0.7
geopandas==0.10.2
gitdb==4.0.9; python_version >= '3.6'
gitpython==3.1.27; python_version >= '3.7'
idna==3.4; python_version >= '3.5'
imageio==2.21.3; python_version >= '3.7'
importlib-metadata==4.12.0; python_version >= '3.7'
importlib-resources==5.9.0; python_version < '3.9'
jinja2==3.1.2; python_version >= '3.7'
jmespath==1.0.1; python_version >= '3.7'
jsonschema==4.16.0; python_version >= '3.7'
kiwisolver==1.4.4; python_version >= '3.7'
markupsafe==2.1.1; python_version >= '3.7'
matplotlib==3.4.2
munch==2.5.0
networkx==2.8.6; python_version >= '3.8'
numpy==1.23.3
onnxruntime==1.11.1
opencv-python-headless==4.1.2.30
packaging==21.3; python_version >= '3.6'
pandas==1.2.2
pillow==9.2.0
pkgutil-resolve-name==1.3.10; python_version < '3.9'
plotly==5.0.0
protobuf==3.20.2; python_version >= '3.7'
pyarrow==9.0.0; python_version >= '3.7'
pycparser==2.21
pydeck==0.8.0b3; python_version >= '3.7'
pygments==2.13.0; python_version >= '3.6'
pympler==1.0.1; python_version >= '3.6'
pyparsing==3.0.9; python_full_version >= '3.6.8'
pyproj==3.4.0; python_version >= '3.8'
pyrsistent==0.18.1; python_version >= '3.7'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==0.21.0
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pywavelets==1.3.0; python_version >= '3.7'
rasterio==1.2.10
rasterstats==0.16.0
requests==2.28.1; python_version >= '3.7' and python_version < '4'
rich==12.5.1; python_full_version >= '3.6.3' and python_version < '4'
s3transfer==0.6.0; python_version >= '3.7'
scikit-image==0.19.3
scipy==1.9.1; python_version < '3.12' and python_version >= '3.8'
semver==2.13.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
setuptools==65.3.0; python_version >= '3.7'
shapely==1.8.4; python_version >= '3.6'
simplejson==3.17.6; python_version >= '2.5' and python_version not in '3.0, 3.1, 3.2'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
smmap==5.0.0; python_version >= '3.6'
snuggs==1.4.7
streamlit==1.10.0
tenacity==8.0.1; python_version >= '3.6'
tifffile==2022.8.12; python_version >= '3.8'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2'
toolz==0.12.0; python_version >= '3.5'
tornado==6.2; python_version >= '3.7'
typing-extensions==4.3.0; python_version >= '3.7'
tzdata==2022.2; python_version >= '3.6'
tzlocal==4.2; python_version >= '3.6'
urllib3==1.26.12; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4'
validators==0.20.0; python_version >= '3.4'
watchdog==2.1.9; platform_system != 'Darwin'
xarray==2022.3.0
zipp==3.8.1; python_version >= '3.7'
